Add percentage dropdown
Added percentage dropdown. This features allows you to calculate a percentage of the weight the user has input. This branch was only meant to include the UI elements but also now includes the logic. Anytime a new percentage is chosen, it automatically recalculates the weights and renders the appropriate UI. Final weight and target weight calculations have also been added and rendered to the user.